Code reuse: routeWrapper
